requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
supabase==2.3.0
pydantic>=2.8.0
python-dotenv==1.0.0
pydub==0.25.1
slowapi==0.1.9
numpy>=1.24.0
torch==2.1.0
torchaudio==2.1.0
soundfile==0.12.1

